requirements.txt
Pillow==11.1.0
qrcode[pil]==8.0
PySide6==6.8.2.1
